Mutable arithmetic

The high level interface can be combined with the low level wrapper to allow for efficient computations using mutable arithmetic.

In the future it would be nice to have an interface to MutableArithmetics.jl, see #118.

The following methods are useful for mutating part of a value

Arblib.radref — Function

radref(x::ArbLike, prec = precision(x))

Return a MagRef referencing the radius of x.

Arblib.midref — Function

midref(x::ArbLike, prec = precision(x))

Return an ArfRef referencing the midpoint of x.

Arblib.realref — Function

realref(z::AcbLike, prec = precision(z))

Return an ArbRef referencing the real part of x.

Arblib.imagref — Function

imagref(z::AcbLike, prec = precision(z))

Return an ArbRef referencing the imaginary part of x.

Arblib.ref — Function

ref(v::Union{ArbVectorOrRef,AcbVectorOrRef}, i)

Similar to v[i] but instead of an Arb or Acb returns an ArbRef or AcbRef which still shares the memory with the i-th entry of v.

ref(A::Union{ArbMatrixOrRef,AcbMatrixOrRef}, i, j)

Similar to A[i,j] but instead of an Arb or Acb returns an ArbRef or AcbRef which still shares the memory with the (i,j)-th entry of A.

ref(p::Union{ArbPoly,ArbSeries,AcbPoly,AcbSeries}, i)

Similar to p[i] but instead of an Arb or Acb returns an ArbRef or AcbRef which still shares the memory with the i-th entry of p[i].

For ArbPoly and AcbPoly it only allows accessing coefficients up to the degree of the polynomial since higher coefficients are not guaranteed to be allocated.

For ArbSeries and AcbSeries it allows accessing coefficients up to the degree of the series. Note that this degree might not be the same as the degree of the underlying polynomial (in case higher order coefficients are zero) but the way they are constructed ensures that the coefficients will always be initialised.

!!! Note: If you use this to change the coefficient in a way so that the degree of the polynomial might change you need to normalise the polynomial afterwards to make sure that Arb recognises the possibly new degree of the polynomial. If the new degree is the same or lower this can be done using Arblib.normalise!. If the new degree is higher you need to manually set the correct value of Arblib.cstruct(p).length to be one higher than the new degree.

Examples

Compare computing $\sqrt{x^2 + y^2}$ using mutable arithmetic with the default.

julia> using Arblib, BenchmarkTools
julia> x = Arb(1 // 3)[0.33333333333333333333333333333333333333333333333333333333333333333333333333333 +/- 4.78e-78]
julia> y = Arb(1 // 5)[0.20000000000000000000000000000000000000000000000000000000000000000000000000000 +/- 3.89e-78]
julia> res = zero(x)0
julia> f(x, y) = sqrt(x^2 + y^2)f (generic function with 1 method)
julia> f!(res, x, y) = begin
           Arblib.sqr!(res, x)
           Arblib.fma!(res, res, y, y)
           return Arblib.sqrt!(res, res)
       endf! (generic function with 1 method)
julia> @benchmark f($x, $y) samples=10000 evals=500BenchmarkTools.Trial: 6075 samples with 500 evaluations.
 Range (min … max):  819.830 ns … 381.562 μs  ┊ GC (min … max):  0.00% … 79.59%
 Time  (median):     898.632 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.637 μs ±  11.005 μs  ┊ GC (mean ± σ):  20.63% ±  3.08%

     ▇█▂                                                        
  ▂▂▂████▆▅▄▃▃▂▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▁▁▁▁▁▁▂▁▁▁▂▂▁▁▂▂▂▁▁▂▂▂▂▁▂▂ ▃
  820 ns           Histogram: frequency by time         1.76 μs <

 Memory estimate: 448 bytes, allocs estimate: 8.
julia> @benchmark f!($res, $x, $y) samples=10000 evals=500BenchmarkTools.Trial: 10000 samples with 500 evaluations.
 Range (min … max):  524.012 ns … 616.814 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     526.214 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   527.545 ns ±   4.754 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▁▄▇███▇▅▃ ▁▁▂▂▂▂▁               ▁▁                           ▂
  ▄█████████████████▇▆▄▅▅▄▅▅▃▄▅▄▅▆█████▇▇▇▆▇▇▇▇▇▇▆▇▇▆▇▇▇▇▇▆▇▆▆▅ █
  524 ns        Histogram: log(frequency) by time        547 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

Set the radius of the real part of an Acb.

julia> using Arblib
julia> z = Acb(1, 2)1.0000000000000000000000000000000000000000000000000000000000000000000000000000 + 2.0000000000000000000000000000000000000000000000000000000000000000000000000000im
julia> Arblib.set!(Arblib.radref(Arblib.realref(z)), 1e-10)(922337204 * 2^-63)
julia> z[1.000000000 +/- 1.01e-10] + 2.0000000000000000000000000000000000000000000000000000000000000000000000000000im

Compare a naive implementation of polynomial evaluation using mutable arithmetic with one not using using it.

julia> using Arblib, BenchmarkTools
julia> p = ArbPoly(1:10)1.0000000000000000000000000000000000000000000000000000000000000000000000000000 + 2.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x + 3.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^2 + 4.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^3 + 5.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^4 + 6.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^5 + 7.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^6 + 8.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^7 + 9.0000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^8 + 10.000000000000000000000000000000000000000000000000000000000000000000000000000⋅x^9
julia> x = Arb(1 // 3)[0.33333333333333333333333333333333333333333333333333333333333333333333333333333 +/- 4.78e-78]
julia> res = zero(x)0
julia> function eval(p, x)
           res = zero(x)
           xi = one(x)
           for i in 0:Arblib.degree(p)
               res += p[i] * xi
               xi *= x
           end
           return res
       endeval (generic function with 2 methods)
julia> function eval!(res, p, x)
           Arblib.zero!(res)
           xi = one(x)
           for i in 0:Arblib.degree(p)
               Arblib.addmul!(res, Arblib.ref(p, i), xi)
               Arblib.mul!(xi, xi, x)
           end
           return res
       endeval! (generic function with 1 method)
julia> @benchmark eval($p, $x) samples = 10000 evals = 30BenchmarkTools.Trial: 10000 samples with 30 evaluations.
 Range (min … max):   4.744 μs …   6.746 ms  ┊ GC (min … max):  0.00% … 79.38%
 Time  (median):      4.944 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   12.568 μs ± 152.519 μs  ┊ GC (mean ± σ):  28.71% ±  2.38%

  ▄██▇▆▅▅▅▄▄▄▃▃▁▂ ▁  ▁▁ ▁        ▂▃▄▃▃▂▂▂ ▁                    ▂
  █████████████████████████▇▇▇▆▇██████████████▆▆▆▇▆▆▆▆▆▅▆▅▅▃▁▅ █
  4.74 μs       Histogram: log(frequency) by time       7.9 μs <

 Memory estimate: 3.94 KiB, allocs estimate: 70.
julia> @benchmark eval!($res, $p, $x) samples = 10000 evals = 30BenchmarkTools.Trial: 10000 samples with 30 evaluations.
 Range (min … max):  2.017 μs …  1.397 ms  ┊ GC (min … max): 0.00% … 40.81%
 Time  (median):     2.077 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.238 μs ± 13.953 μs  ┊ GC (mean ± σ):  2.55% ±  0.41%

   ▄█▂            ▂                                          
  ▂███▆▆▆▅▄▂▃▂▂▂▃▇██▆▃▄▃▃▃▃▂▂▂▁▁▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  2.02 μs        Histogram: frequency by time        2.38 μs <

 Memory estimate: 160 bytes, allocs estimate: 3.
julia> @benchmark $p($x) samples = 10000 evals = 30 # Arb implementation for referenceBenchmarkTools.Trial: 10000 samples with 30 evaluations.
 Range (min … max):  1.577 μs …  1.395 ms  ┊ GC (min … max): 0.00% … 41.01%
 Time  (median):     1.643 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.802 μs ± 13.930 μs  ┊ GC (mean ± σ):  3.17% ±  0.41%

    █▃▄▁            ▁ ▂                                      
  ▁▂████▅▇▆▃▅▅▂▃▃▂▂▃█▆██▄▅▄▂▃▄▃▂▂▂▁▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  1.58 μs        Histogram: frequency by time         1.9 μs <

 Memory estimate: 160 bytes, allocs estimate: 3.