Types

The package defines a number of types for the high level interface.

Basic

These types directly map to corresponding Arb types.

Arblib.Mag — Type

Mag <: Real

Arblib.Arf — Type

Arf <: AbstractFloat

Arblib.Arb — Type

Arb <: AbstractFloat

Arblib.Acb — Type

Acb <: Number

Arblib.ArbVector — Type

ArbVector <: DenseVector{Arb}

Arblib.AcbVector — Type

AcbVector <: DenseVector{Acb}

Arblib.ArbPoly — Type

ArbPoly

Arblib.AcbPoly — Type

AcbPoly

Arblib.ArbMatrix — Type

ArbMatrix <: DenseMatrix{Arb}

Arblib.AcbMatrix — Type

AcbMatrix <: DenseMatrix{Acb}

Series

The package defines two series types, which are wrapper for the polynomial types with a specified degree.

Arblib.ArbSeries — Type

ArbSeries <: Number

This type should be considered experimental, the interface for it might change in the future.

Arblib.AcbSeries — Type

AcbSeries <: Number

This type should be considered experimental, the interface for it might change in the future.

Ref

In addition to these there are a number of Ref types, which allow for non-allocating access in a number of cases.

Arblib.MagRef — Type

MagRef <: Real

Arblib.ArfRef — Type

ArfRef <: AbstractFloat

Arblib.ArbRef — Type

ArbRef <: AbstractFloat

Arblib.AcbRef — Type

AcbRef <: Number

Arblib.ArbRefVector — Type

ArbRefVector <: DenseMatrix{ArbRef}

Arblib.AcbRefVector — Type

AcbRefVector <: DenseMatrix{AcbRef}

Arblib.ArbRefMatrix — Type

ArbRefMatrix <: DenseMatrix{ArbRef}

Arblib.AcbRefMatrix — Type

AcbRefMatrix <: DenseMatrix{AcbRef}

Correspondence between types

We have the following table for the correspondence with between the Low level wrapper types and the high level interface types.

Arb	Wrapper	High level	Ref

mag_t	mag_struct	Mag	MagRef

arf_t	arf_struct	Arf	ArfRef

arb_t	arb_struct	Arb	ArbRef

acb_t	acb_struct	Acb	AcbRef

arb_t*	arb_vec_struct	ArbVector	ArbRefVector

acb_t*	acb_vec_struct	AcbVector	AcbRefVector

arb_poly_t	arb_poly_struct	ArbPoly or ArbSeries

acb_poly_t	acb_poly_struct	AcbPoly or AcbSeries

arb_mat_t	arb_mat_struct	ArbMatrix	ArbRefMatrix

acb_mat_t	acb_mat_struct	AcbMatrix	AcbRefMatrix