Types

The package defines the following types which map directly to Arb types with the corresponding name. In most cases you should not use these types directly but use the types from the High level interface.

Arblib.mag_struct — Type

mag_struct

Arblib.arf_struct — Type

arf_struct

Arblib.arb_struct — Type

arb_struct

Arblib.acb_struct — Type

acb_struct

Arblib.arb_vec_struct — Type

arb_vec_struct

Arblib.acb_vec_struct — Type

acb_vec_struct

Arblib.arb_poly_struct — Type

arb_poly_struct

Arblib.acb_poly_struct — Type

acb_poly_struct

Arblib.arb_mat_struct — Type

arb_mat_struct

Arblib.acb_mat_struct — Type

acb_mat_struct

Arblib.calc_integrate_opt_struct — Type

calc_integrate_opt_struct

For each low-level type there is a union of types that can be interpreted as the low-level type. These are the types that can be given directly as arguments to the low-level methods. Below you find these union-types.

Arblib.MagLike — Type

MagLike = Union{Mag,MagRef,mag_struct,Ptr{mag_struct}}

Arblib.ArfLike — Type

ArfLike = Union{Arf,ArfRef,arf_struct,Ptr{arf_struct}}}

Arblib.ArbLike — Type

ArbLike = Union{Arb,ArbRef,arb_struct,Ptr{arb_struct}}}

Arblib.AcbLike — Type

AcbLike = Union{Acb,AcbRef,acb_struct,Ptr{acb_struct}}}

Arblib.ArbVectorLike — Type

ArbVectorLike = Union{ArbVector,ArbRefVector,arb_vec_struct}

Arblib.AcbVectorLike — Type

AcbVectorLike = Union{AcbVector,AcbRefVector,acb_vec_struct}

Arblib.ArbMatrixLike — Type

ArbMatrixLike = Union{ArbMatrix,ArbRefMatrix,arb_mat_struct)}

Arblib.AcbMatrixLike — Type

AcbMatrixLike = Union{AcbMatrix,AcbRefMatrix,acb_mat_struct}

Arblib.ArbPolyLike — Type

ArbPolyLike = Union{ArbPoly,ArbSeries,arb_poly_struct}

Arblib.AcbPolyLike — Type

AcbPolyLike = Union{AcbPoly,AcbSeries,acb_poly_struct}

Note that the show method is overloaded for these union types, this is to make method declarations easier to read in the REPL. As an example we can print the methods for Arblib.sin! and we see that it prints ArbLike instead of the much longer Union{Ptr{Arblib.arb_struct}, Arb, ArbRef, Arblib.arb_struct}.

julia> using Arblib
julia> methods(Arblib.sin!)# 4 methods for generic function "sin!" from Arblib:
 [1] sin!(s::AcbLike, z::AcbLike; prec)
     @ ~/work/Arblib.jl/Arblib.jl/src/ArbCall/ArbFunction.jl:176
 [2] sin!(s::AcbLike, z::AcbLike, prec::Integer)
     @ ~/work/Arblib.jl/Arblib.jl/src/ArbCall/ArbFunction.jl:152
 [3] sin!(s::ArbLike, x::ArbLike; prec)
     @ ~/work/Arblib.jl/Arblib.jl/src/ArbCall/ArbFunction.jl:176
 [4] sin!(s::ArbLike, x::ArbLike, prec::Integer)
     @ ~/work/Arblib.jl/Arblib.jl/src/ArbCall/ArbFunction.jl:152